Added check to ensure listing is a punk
If the title does not contain a punk ID, then we cannot verify that it's legit, because we can't cross-reference the punk ID in the bitcoinpunks.com API.
If the title contains other numbers, they will be incorrectly parsed as the punk ID and return a false negative. 
As a result, we check to ensure that the title contains "punk", so we can ensure that any number being parsed is a punk ID. This is a heuristic that works in most cases. 
If we cannot reliably parse a punk ID, we link the user to the manual verification method on twitter, so they can DYOR and determine for themselves.